+++ src/main/java/com/project/hrbank/backup/provider/EmployeesLogCsvFileProvider.java
@@ -19,8 +19,8 @@ import org.springframework.stereotype.Component;
 
 import com.project.hrbank.entity.Employee;
 import com.project.hrbank.entity.EmployeeLogs;
-import com.project.hrbank.file.entity.FileEntity;
-import com.project.hrbank.file.entity.enums.FileExtension;
+import com.project.hrbank.entity.FileEntity;
+import com.project.hrbank.entity.enums.FileExtension;
 import com.project.hrbank.repository.EmployeeRepository;
 
 @Component
