+++ src/main/java/com/project/hrbank/backup/provider/LogFileProvider.java
@@ -14,7 +14,7 @@ import java.time.format.DateTimeFormatter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.project.hrbank.file.entity.enums.FileExtension;
+import com.project.hrbank.entity.enums.FileExtension;
 
 @Component
 public class LogFileProvider {
