+++ src/main/java/com/project/hrbank/config/file/FileHandlerConfig.java
@@ -1,12 +1,12 @@
-package com.project.hrbank.file.config;
+package com.project.hrbank.config.file;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.project.hrbank.file.handler.FileHandler;
-import com.project.hrbank.file.handler.impl.CsvFileHandler;
-import com.project.hrbank.file.handler.impl.ImageFileHandler;
-import com.project.hrbank.file.handler.impl.LogFileHandler;
+import com.project.hrbank.util.handler.FileHandler;
+import com.project.hrbank.util.handler.impl.CsvFileHandler;
+import com.project.hrbank.util.handler.impl.ImageFileHandler;
+import com.project.hrbank.util.handler.impl.LogFileHandler;
 
 @Configuration
 public class FileHandlerConfig {
