+++ src/main/java/com/project/hrbank/config/file/FileUploadConfig.java
@@ -1,4 +1,4 @@
-package com.project.hrbank.file.config;
+package com.project.hrbank.config.file;
 
 import org.springframework.boot.web.servlet.MultipartConfigFactory;
 import org.springframework.context.annotation.Bean;
