+++ src/main/java/com/project/hrbank/repository/FileRepository.java
@@ -1,9 +1,9 @@
-package com.project.hrbank.file.repository;
+package com.project.hrbank.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.project.hrbank.file.entity.FileEntity;
+import com.project.hrbank.entity.FileEntity;
 
 @Repository
 public interface FileRepository extends JpaRepository<FileEntity, Long> {
